Split crates and workflows by ownership
Turn rapport-command, rapport-files, and rapport-git roots into outlines; split Policy Context into CLI, rendering, doctor, signoff, and mutation modules; split Work into CLI, status, checkpoint, rebase, and orchestration modules; preserve all workspace tests.
