+++ crates/rapport-files/src/lib.rs
@@ -1,579 +1,16 @@
 //! Testable filesystem primitives for Rapport workflow code.
+//!
+//! The crate exposes one filesystem contract with production and in-memory
+//! implementations owned by focused modules.
+
+mod file_system;
+mod memory;
+mod real;
 
 pub use camino::{Utf8Path, Utf8PathBuf};
-use std::collections::{BTreeSet, HashMap, HashSet};
-use std::io::{self, Write as _};
-
-pub trait FileSystem {
-    fn is_dir(&self, path: impl AsRef<Utf8Path>) -> bool;
-
-    fn is_file(&self, path: impl AsRef<Utf8Path>) -> bool;
-
-    /// Resolve symlinks and return an absolute canonical path.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error when the path does not exist or cannot be canonicalized.
-    fn canonicalize(&self, path: impl AsRef<Utf8Path>) -> io::Result<Utf8PathBuf> {
-        let path = path.as_ref();
-        if self.exists(path) {
-            Ok(path.to_path_buf())
-        } else {
-            Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                format!("{path} not found"),
-            ))
-        }
-    }
-
-    /// Return the Git file mode for a regular working-tree file.
-    ///
-    /// In-memory and non-Unix implementations default to a non-executable
-    /// regular file. Production Unix filesystems preserve the executable bit.
-    ///
-    /// # Errors
-    ///
-    /// Returns [`io::ErrorKind::NotFound`] when the file does not exist.
-    fn git_file_mode(&self, path: impl AsRef<Utf8Path>) -> io::Result<u32> {
-        let path = path.as_ref();
-        if self.is_file(path) {
-            Ok(0o100_644)
-        } else {
-            Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                format!("{path} not found"),
-            ))
-        }
-    }
-
-    /// Read a UTF-8 file from the filesystem.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the path cannot be read.
-    fn read_to_string(&self, path: impl AsRef<Utf8Path>) -> io::Result<String>;
-
-    /// Read arbitrary file bytes from the filesystem.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the path cannot be read.
-    fn read_bytes(&self, path: impl AsRef<Utf8Path>) -> io::Result<Vec<u8>> {
-        self.read_to_string(path).map(String::into_bytes)
-    }
-
-    /// Read the immediate children of a directory.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the directory cannot be read.
-    fn read_dir(&self, path: impl AsRef<Utf8Path>) -> io::Result<Vec<Utf8PathBuf>>;
-
-    /// Create a directory and all missing parent directories.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the directory cannot be created.
-    fn create_dir_all(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()>;
-
-    /// Write a UTF-8 file, replacing any existing contents.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the path cannot be written.
-    fn write_string(
-        &mut self,
-        path: impl AsRef<Utf8Path>,
-        contents: impl AsRef<str>,
-    ) -> io::Result<()>;
-
-    /// Append one UTF-8 line to a file, creating it when it does not exist.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the path cannot be appended.
-    fn append_line(&mut self, path: impl AsRef<Utf8Path>, line: impl AsRef<str>) -> io::Result<()>;
-
-    /// Remove a file from the filesystem.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the file cannot be removed.
-    fn remove_file(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()>;
-
-    /// Atomically rename a file or directory within one filesystem.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the source cannot be
-    /// renamed or the destination already exists.
-    fn rename(&mut self, from: impl AsRef<Utf8Path>, to: impl AsRef<Utf8Path>) -> io::Result<()>;
-
-    /// Remove a directory and every descendant.
-    ///
-    /// # Errors
-    ///
-    /// Returns the underlying filesystem error when the directory cannot be
-    /// removed.
-    fn remove_dir_all(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()>;
-
-    fn exists(&self, path: impl AsRef<Utf8Path>) -> bool {
-        self.is_dir(path.as_ref()) || self.is_file(path)
-    }
-}
-
-#[derive(Debug, Default, Clone, Copy)]
-pub struct RealFileSystem;
-
-impl FileSystem for RealFileSystem {
-    fn is_dir(&self, path: impl AsRef<Utf8Path>) -> bool {
-        path.as_ref().is_dir()
-    }
-
-    fn is_file(&self, path: impl AsRef<Utf8Path>) -> bool {
-        path.as_ref().is_file()
-    }
-
-    fn canonicalize(&self, path: impl AsRef<Utf8Path>) -> io::Result<Utf8PathBuf> {
-        let path = std::fs::canonicalize(path.as_ref())?;
-        Utf8PathBuf::from_path_buf(path).map_err(|path| {
-            io::Error::new(
-                io::ErrorKind::InvalidData,
-                format!("canonical path is not UTF-8: {}", path.display()),
-            )
-        })
-    }
-
-    fn git_file_mode(&self, path: impl AsRef<Utf8Path>) -> io::Result<u32> {
-        let metadata = std::fs::metadata(path.as_ref())?;
-        if !metadata.is_file() {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("{} is not a regular file", path.as_ref()),
-            ));
-        }
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt as _;
-            if metadata.permissions().mode() & 0o111 != 0 {
-                return Ok(0o100_755);
-            }
-        }
-        Ok(0o100_644)
-    }
-
-    fn read_to_string(&self, path: impl AsRef<Utf8Path>) -> io::Result<String> {
-        std::fs::read_to_string(path.as_ref())
-    }
-
-    fn read_bytes(&self, path: impl AsRef<Utf8Path>) -> io::Result<Vec<u8>> {
-        std::fs::read(path.as_ref())
-    }
-
-    fn read_dir(&self, path: impl AsRef<Utf8Path>) -> io::Result<Vec<Utf8PathBuf>> {
-        let mut entries = Vec::new();
-        for entry in std::fs::read_dir(path.as_ref())? {
-            let entry = entry?;
-            let path = Utf8PathBuf::from_path_buf(entry.path()).map_err(|path| {
-                io::Error::new(
-                    io::ErrorKind::InvalidData,
-                    format!("{} is not a UTF-8 path", path.display()),
-                )
-            })?;
-            entries.push(path);
-        }
-        entries.sort();
-        Ok(entries)
-    }
-
-    fn create_dir_all(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()> {
-        std::fs::create_dir_all(path.as_ref())
-    }
-
-    fn write_string(
-        &mut self,
-        path: impl AsRef<Utf8Path>,
-        contents: impl AsRef<str>,
-    ) -> io::Result<()> {
-        if let Some(parent) = path.as_ref().parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-        std::fs::write(path.as_ref(), contents.as_ref())
-    }
-
-    fn append_line(&mut self, path: impl AsRef<Utf8Path>, line: impl AsRef<str>) -> io::Result<()> {
-        if let Some(parent) = path.as_ref().parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-        let mut file = std::fs::OpenOptions::new()
-            .create(true)
-            .append(true)
-            .open(path.as_ref())?;
-        writeln!(file, "{}", line.as_ref())
-    }
-
-    fn remove_file(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()> {
-        std::fs::remove_file(path.as_ref())
-    }
-
-    fn rename(&mut self, from: impl AsRef<Utf8Path>, to: impl AsRef<Utf8Path>) -> io::Result<()> {
-        std::fs::rename(from.as_ref(), to.as_ref())
-    }
-
-    fn remove_dir_all(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()> {
-        std::fs::remove_dir_all(path.as_ref())
-    }
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct InMemoryFileSystem {
-    directories: HashSet<Utf8PathBuf>,
-    files: HashMap<Utf8PathBuf, String>,
-}
-
-impl InMemoryFileSystem {
-    pub fn add_directory(&mut self, path: impl AsRef<Utf8Path>) {
-        let path = path.as_ref();
-        self.add_parent_directories(path);
-        self.directories.insert(path.to_path_buf());
-    }
-
-    pub fn add_file(&mut self, path: impl AsRef<Utf8Path>) {
-        self.add_file_with_contents(path, "");
-    }
-
-    pub fn add_file_with_contents(
-        &mut self,
-        path: impl AsRef<Utf8Path>,
-        contents: impl Into<String>,
-    ) {
-        if let Some(parent) = path.as_ref().parent() {
-            self.add_directory(parent);
-        }
-        self.files
-            .insert(path.as_ref().to_path_buf(), contents.into());
-    }
-
-    fn add_parent_directories(&mut self, path: &Utf8Path) {
-        let mut current = path.parent();
-        while let Some(parent) = current {
-            if parent.as_str().is_empty() {
-                break;
-            }
-            self.directories.insert(parent.to_path_buf());
-            current = parent.parent();
-        }
-    }
-}
-
-impl FileSystem for InMemoryFileSystem {
-    fn is_dir(&self, path: impl AsRef<Utf8Path>) -> bool {
-        self.directories.contains(path.as_ref())
-    }
-
-    fn is_file(&self, path: impl AsRef<Utf8Path>) -> bool {
-        self.files.contains_key(path.as_ref())
-    }
-
-    fn read_to_string(&self, path: impl AsRef<Utf8Path>) -> io::Result<String> {
-        self.files.get(path.as_ref()).cloned().ok_or_else(|| {
-            io::Error::new(
-                io::ErrorKind::NotFound,
-                format!("{} not found", path.as_ref()),
-            )
-        })
-    }
-
-    fn read_dir(&self, path: impl AsRef<Utf8Path>) -> io::Result<Vec<Utf8PathBuf>> {
-        let path = path.as_ref();
-        if !self.is_dir(path) {
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                format!("{path} not found"),
-            ));
-        }
-
-        let mut entries = BTreeSet::new();
-        for directory in &self.directories {
-            if directory.parent() == Some(path) {
-                entries.insert(directory.clone());
-            }
-        }
-        for file in self.files.keys() {
-            if file.parent() == Some(path) {
-                entries.insert(file.clone());
-            }
-        }
-        Ok(entries.into_iter().collect())
-    }
-
-    fn create_dir_all(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()> {
-        self.add_directory(path);
-        Ok(())
-    }
-
-    fn write_string(
-        &mut self,
-        path: impl AsRef<Utf8Path>,
-        contents: impl AsRef<str>,
-    ) -> io::Result<()> {
-        self.add_file_with_contents(path, contents.as_ref());
-        Ok(())
-    }
-
-    fn append_line(&mut self, path: impl AsRef<Utf8Path>, line: impl AsRef<str>) -> io::Result<()> {
-        if let Some(parent) = path.as_ref().parent() {
-            self.add_directory(parent);
-        }
-        let entry = self.files.entry(path.as_ref().to_path_buf()).or_default();
-        entry.push_str(line.as_ref());
-        entry.push('\n');
-        Ok(())
-    }
-
-    fn remove_file(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()> {
-        self.files.remove(path.as_ref()).map_or_else(
-            || {
-                Err(io::Error::new(
-                    io::ErrorKind::NotFound,
-                    format!("{} not found", path.as_ref()),
-                ))
-            },
-            |_| Ok(()),
-        )
-    }
-
-    fn rename(&mut self, from: impl AsRef<Utf8Path>, to: impl AsRef<Utf8Path>) -> io::Result<()> {
-        let from = from.as_ref().to_path_buf();
-        let to = to.as_ref().to_path_buf();
-        if self.exists(&to) {
-            return Err(io::Error::new(
-                io::ErrorKind::AlreadyExists,
-                format!("{to} already exists"),
-            ));
-        }
-        if let Some(contents) = self.files.remove(&from) {
-            if let Some(parent) = to.parent() {
-                self.add_directory(parent);
-            }
-            self.files.insert(to, contents);
-            return Ok(());
-        }
-        if !self.directories.contains(&from) {
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                format!("{from} not found"),
-            ));
-        }
-
-        let directories = self
-            .directories
-            .iter()
-            .filter(|path| **path == from || path.starts_with(&from))
-            .cloned()
-            .collect::<Vec<_>>();
-        let files = self
-            .files
-            .iter()
-            .filter(|(path, _)| path.starts_with(&from))
-            .map(|(path, contents)| (path.clone(), contents.clone()))
-            .collect::<Vec<_>>();
-        self.directories
-            .retain(|path| *path != from && !path.starts_with(&from));
-        self.files.retain(|path, _| !path.starts_with(&from));
-        if let Some(parent) = to.parent() {
-            self.add_directory(parent);
-        }
-        for path in directories {
-            let relative = path.strip_prefix(&from).map_err(io::Error::other)?;
-            self.directories.insert(to.join(relative));
-        }
-        for (path, contents) in files {
-            let relative = path.strip_prefix(&from).map_err(io::Error::other)?;
-            self.files.insert(to.join(relative), contents);
-        }
-        Ok(())
-    }
-
-    fn remove_dir_all(&mut self, path: impl AsRef<Utf8Path>) -> io::Result<()> {
-        let path = path.as_ref();
-        if !self.directories.contains(path) {
-            return Err(io::Error::new(
-                io::ErrorKind::NotFound,
-                format!("{path} not found"),
-            ));
-        }
-        self.files.retain(|file, _| !file.starts_with(path));
-        self.directories
-            .retain(|directory| directory != path && !directory.starts_with(path));
-        Ok(())
-    }
-}
+pub use file_system::FileSystem;
+pub use memory::InMemoryFileSystem;
+pub use real::RealFileSystem;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use claims::{assert_err, assert_ok};
-
-    #[test]
-    fn in_memory_file_system_recognizes_added_directory() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_directory("/work");
-
-        assert!(fs.is_dir("/work"));
-    }
-
-    #[test]
-    fn in_memory_file_system_does_not_treat_directory_as_file() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_directory("/work");
-
-        assert!(!fs.is_file("/work"));
-    }
-
-    #[test]
-    fn in_memory_file_system_recognizes_added_file() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_file("/work/Cargo.toml");
-
-        assert!(fs.is_file("/work/Cargo.toml"));
-    }
-
-    #[test]
-    fn in_memory_file_system_does_not_treat_file_as_directory() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_file("/work/Cargo.toml");
-
-        assert!(!fs.is_dir("/work/Cargo.toml"));
-    }
-
-    #[test]
-    fn in_memory_file_system_exists_recognizes_added_file() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_file("/work/Cargo.toml");
-
-        assert!(fs.exists("/work/Cargo.toml"));
-    }
-
-    #[test]
-    fn in_memory_file_system_reads_added_files() {
-        let mut fs = InMemoryFileSystem::default();
-        let path = Utf8PathBuf::from("/work/Package.swift");
-        fs.add_file_with_contents(&path, "// swift-tools-version: 6.0\n");
-
-        assert_eq!(
-            assert_ok!(fs.read_to_string(&path)),
-            "// swift-tools-version: 6.0\n"
-        );
-    }
-
-    #[test]
-    fn in_memory_file_system_reports_missing_files() {
-        let fs = InMemoryFileSystem::default();
-        let path = Utf8PathBuf::from("/work/Package.swift");
-
-        let err = assert_err!(fs.read_to_string(&path));
-
-        assert_eq!(err.kind(), io::ErrorKind::NotFound);
-    }
-
-    #[test]
-    fn in_memory_file_system_reads_immediate_children_sorted() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_file("/work/z.toml");
-        fs.add_file("/work/a.toml");
-        fs.add_file("/work/nested/config.toml");
-
-        let children = assert_ok!(fs.read_dir("/work"));
-
-        assert_eq!(
-            children,
-            vec![
-                Utf8PathBuf::from("/work/a.toml"),
-                Utf8PathBuf::from("/work/nested"),
-                Utf8PathBuf::from("/work/z.toml"),
-            ]
-        );
-    }
-
-    #[test]
-    fn in_memory_file_system_writes_files() {
-        let mut fs = InMemoryFileSystem::default();
-
-        assert_ok!(fs.write_string("/work/.rapport/work.toml", "schema_version = 1\n"));
-
-        assert!(fs.is_dir("/work/.rapport"));
-        assert_eq!(
-            assert_ok!(fs.read_to_string("/work/.rapport/work.toml")),
-            "schema_version = 1\n"
-        );
-    }
-
-    #[test]
-    fn in_memory_file_system_appends_lines() {
-        let mut fs = InMemoryFileSystem::default();
-
-        assert_ok!(fs.append_line("/work/.rapport/events.jsonl", "{\"one\":1}"));
-        assert_ok!(fs.append_line("/work/.rapport/events.jsonl", "{\"two\":2}"));
-
-        assert_eq!(
-            assert_ok!(fs.read_to_string("/work/.rapport/events.jsonl")),
-            "{\"one\":1}\n{\"two\":2}\n"
-        );
-    }
-
-    #[test]
-    fn in_memory_file_system_removes_files() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_file("/work/.rapport/work.toml");
-
-        assert_ok!(fs.remove_file("/work/.rapport/work.toml"));
-
-        assert!(!fs.is_file("/work/.rapport/work.toml"));
-    }
-
-    #[test]
-    fn rename_should_move_a_complete_directory_tree() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_file_with_contents("/history/pending/work.toml", "version = 2\n");
-        fs.add_file_with_contents("/history/pending/tasks/TASK_001.toml", "version = 1\n");
-
-        assert_ok!(fs.rename("/history/pending", "/history/019f53"));
-
-        assert!(
-            !fs.exists("/history/pending"),
-            "expecting atomic rename to remove the pending tree"
-        );
-        assert_eq!(
-            assert_ok!(fs.read_to_string("/history/019f53/work.toml")),
-            "version = 2\n",
-            "expecting atomic rename to preserve file contents"
-        );
-        assert!(
-            fs.is_file("/history/019f53/tasks/TASK_001.toml"),
-            "expecting atomic rename to preserve nested Task files"
-        );
-    }
-
-    #[test]
-    fn remove_dir_all_should_remove_every_descendant() {
-        let mut fs = InMemoryFileSystem::default();
-        fs.add_file("/history/019f53/work.toml");
-        fs.add_file("/history/019f53/tasks/TASK_001.toml");
-
-        assert_ok!(fs.remove_dir_all("/history/019f53"));
-
-        assert!(
-            !fs.exists("/history/019f53"),
-            "expecting recursive removal to delete the selected record"
-        );
-        assert!(
-            !fs.exists("/history/019f53/tasks/TASK_001.toml"),
-            "expecting recursive removal to delete nested Task files"
-        );
-        assert!(
-            fs.is_dir("/history"),
-            "expecting recursive removal to preserve the parent history directory"
-        );
-    }
-}
+mod tests;
