+++ crates/rapport/src/policy_context/mod.rs
@@ -1,16 +1,22 @@
 //! Contextual repository policy.
 
 mod boundary;
+mod cli;
 mod command;
+mod doctor;
 mod domain;
 mod error;
+mod render;
 mod repository;
+mod signoff;
 mod workflow;
 
 #[cfg(test)]
 mod tests;
 
-pub(crate) use command::{Cli, doctor_all, run};
+pub(crate) use cli::Cli;
+pub(crate) use command::run;
+pub(crate) use doctor::doctor_all;
 pub(crate) use error::Error;
 pub(crate) use workflow::{SHARED_PATH as SHARED_WORKFLOW_PATH, write_shared};
 
@@ -112,7 +118,7 @@ pub(crate) fn effective_policy_digest_for_paths<'path>(
     let mut rendered = paths
         .into_iter()
         .map(|path| {
-            command::show(fs, repo_root, path, false).map(|policy| (path.to_string(), policy))
+            render::show(fs, repo_root, path, false).map(|policy| (path.to_string(), policy))
         })
         .collect::<Result<Vec<_>, _>>()?;
     rendered.sort_by(|left, right| left.0.cmp(&right.0));
